voice-memos v1

+++ src/index.ts
@@ -2,7 +2,8 @@ import express from 'express';
 import multer from 'multer'
 import { uploadS3 } from './s3/uploadToBucket';
 import fs from 'node:fs'
-import { run } from './transcribe/start';
+import { transcribe } from './transcribe/start';
+import { env } from './env/env';
 
 const app = express()
 const upload = multer({dest: 'tmp/'});
@@ -12,12 +13,17 @@ app.get('/', (req,res)=>{
 
 app.post('/transcribe',upload.single('file'), async (req:any, res)=>{
   const filePath = req.file.path;
-  const fileName = req.file.originalname
-  
- await uploadS3('voice-memos-bucket', fileName, filePath);
- fs.unlinkSync(filePath)
- await run('JobNodeBack', fileName)
-  res.send(200);
+  const fileName = req.file.originalname;
+  const jobName = req.body.jobName;
+  if (!jobName){
+    res.send('jobName must be sended in the request body. Please try again!').sendStatus(400)
+    throw new Error('jobName missing.')
+  }
+  await uploadS3(env.BUCKET_NAME, fileName, filePath);
+  await transcribe(fileName, jobName);
+  fs.unlinkSync(filePath);
+ const _url = env.BASE_URL + `/${jobName}.json`
+res.send(`Job started! When finished, available at ${_url}`).sendStatus(200);
 })
 
 app.listen(3030, ()=>{
